+++ src/components/dashboards/Admin-dashboard.tsx
@@ -3,49 +3,23 @@
 import SearchRequestsFilterPanel from "@/components/search-requests-filter-panel/Search-requests-filter-panel";
 import Spinner from "@/components/spinner/Spinner";
 import { formatPhoneNumber } from "@/lib/utils/phone-number/format-phone-number";
+import { getTimeSlotHours } from "@/lib/utils/time-slot/time-slot";
 import { AllUserRequestsAdminGetResponse } from "@/types/api-responses/admin-resident-requests-api-response";
+import { TimeSlot } from "@/types/time-slot";
 import { RequestStatus } from "@prisma/client";
 import { ResidentRequestService } from "app/services/resident-request-service";
 import { useSession } from "next-auth/react";
-import Link from "next/link";
 import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
-import FormattedTimeSlotDateTime from "../formatted-time-slot-date-time.tsx/FormattedTimeSlotDateTime";
+import { useEffect, useRef, useState } from "react";
 import { toast, ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
+import FormattedTimeSlotDateTime from "../formatted-time-slot-date-time.tsx/FormattedTimeSlotDateTime";
+import RouteManager from "../route-manager/Route-manager";
+import { ActionButtons } from "./resources/Action-buttons";
+import { DaySection } from "./resources/Day-section";
+import { statusToColorMap } from "./resources/status-color-map";
 
 const MAX_TAKE = 9; // max cards on page
-
-const statusToColorMap: Record<
-  RequestStatus,
-  { border: string; bg: string; text: string }
-> = {
-  COMPLETED: {
-    border: "border-green-500",
-    bg: "bg-green-50",
-    text: "text-green-600",
-  },
-  PENDING: {
-    border: "border-yellow-500",
-    bg: "bg-yellow-50",
-    text: "text-yellow-600",
-  },
-  CANCELED: {
-    border: "border-red-500",
-    bg: "bg-red-50",
-    text: "text-red-600",
-  },
-};
-
-const formatDate = (dateString: string | Date) => {
-  const date =
-    typeof dateString === "string" ? new Date(dateString) : dateString;
-  return date.toLocaleDateString("en-US", {
-    month: "short",
-    day: "numeric",
-    year: "numeric",
-  });
-};
 
 export default function AdminDashboard() {
   const router = useRouter();
@@ -59,11 +33,13 @@ export default function AdminDashboard() {
     AllUserRequestsAdminGetResponse[]
   >([]);
   const [isSearching, setIsSearch] = useState(false);
-  const [date, setDate] = useState<string | null>(null);
-
   const [selectedRequest, setSelectedRequest] =
     useState<AllUserRequestsAdminGetResponse | null>(null);
   const [isModalClosing, setIsModalClosing] = useState(false);
+  const [isRouteOptimizedModalOpen, setIsRouteOptimizedModalOpen] =
+    useState(true);
+
+  const date = useRef<string | null>(null);
 
   useEffect(() => {
     if (status === "unauthenticated") {
@@ -76,16 +52,16 @@ export default function AdminDashboard() {
   }, [status, session?.user?.isAdmin, router]);
 
   useEffect(() => {
-    if (session?.user?.isAdmin && !isSearching && !date) {
+    if (session?.user?.isAdmin && !isSearching) {
       fetchAllRequests();
     }
-  }, [session?.user?.isAdmin, pageNumber, isSearching, date]);
+  }, [session?.user?.isAdmin, pageNumber, isSearching]);
 
   useEffect(() => {
-    if (session?.user?.isAdmin && date) {
+    if (session?.user?.isAdmin) {
       fetchAllRequestsByDate();
     }
-  }, [session?.user?.isAdmin, date]);
+  }, [session?.user?.isAdmin]);
 
   const fetchAllRequests = async () => {
     try {
@@ -112,19 +88,20 @@ export default function AdminDashboard() {
   };
 
   const fetchAllRequestsByDate = async () => {
-    if (!date) return;
     try {
       setLoading(true);
       const requests = await ResidentRequestService.adminGetAllRequests({
-        date,
+        date: date.current,
         take: 999,
         skip: 0,
       });
+
       const sorted = requests.residentRequests.sort(
         (a, b) =>
           new Date(a.requestedTimeSlot.startTime).getTime() -
           new Date(b.requestedTimeSlot.startTime).getTime()
       );
+
       setUserRequests(sorted);
       setTotalCount(sorted.length);
       setLoading(false);
@@ -140,7 +117,7 @@ export default function AdminDashboard() {
     filters: Record<string, boolean>,
     selectedDate: string | null
   ) => {
-    setDate(selectedDate!);
+    date.current = selectedDate;
     setIsSearch(true);
 
     try {
@@ -185,7 +162,7 @@ export default function AdminDashboard() {
   };
 
   const handlePagination = (direction: "forward" | "backward") => {
-    if (date) return; // when selecting a date, don't show pagination
+    if (date.current) return; // when selecting a date, don't show pagination
     if (direction === "forward") {
       if ((pageNumber + 1) * MAX_TAKE > totalCount) return;
       setPageNumber((prev) => prev + 1);
@@ -206,7 +183,7 @@ export default function AdminDashboard() {
         requestStatus
       );
       toast.success(`Request updated to ${requestStatus}`);
-      if (date) {
+      if (date.current) {
         await fetchAllRequestsByDate();
       } else {
         await fetchAllRequests();
@@ -222,7 +199,7 @@ export default function AdminDashboard() {
 
   const handleCancelSearch = async () => {
     setIsSearch(false);
-    setDate(null);
+    date.current = null;
     setPageNumber(0);
     await fetchAllRequests();
   };
@@ -255,28 +232,19 @@ export default function AdminDashboard() {
   let noonRequests: AllUserRequestsAdminGetResponse[] = [];
   let eveningRequests: AllUserRequestsAdminGetResponse[] = [];
 
-  if (date) {
+  if (date.current) {
     userRequests.forEach((req) => {
-      const startTimeValue = req.requestedTimeSlot.startTime;
-      const startTimeString =
-        typeof startTimeValue === "string"
-          ? startTimeValue
-          : startTimeValue.toISOString();
+      const startTimeValue = new Date(req.requestedTimeSlot.startTime);
+      const hour = startTimeValue.getHours();
 
-      const hour = parseInt(startTimeString.substring(11, 13), 10);
-
-      if (hour == 11) {
+      if (hour === getTimeSlotHours(TimeSlot.Morning)[0]) {
         morningRequests.push(req);
-      } else if (hour == 14) {
+      } else if (hour === getTimeSlotHours(TimeSlot.Daytime)[0]) {
         noonRequests.push(req);
-      } else if (hour == 17) {
+      } else if (hour === getTimeSlotHours(TimeSlot.Evening)[0]) {
         eveningRequests.push(req);
       }
     });
-
-    morningRequests = morningRequests.slice(0, 3);
-    noonRequests = noonRequests.slice(0, 3);
-    eveningRequests = eveningRequests.slice(0, 3);
   }
 
   return (
@@ -287,12 +255,17 @@ export default function AdminDashboard() {
       </h2>
       <div className="flex justify-between items-center flex-wrap gap-2 mb-8">
         <SearchRequestsFilterPanel onSearch={handleSearch} />
-        <Link
-          className="text-white bg-[#1e3a89] p-2 rounded-md self-center"
-          href="/dashboard/admin/route-manager"
-        >
-          Route Manager
-        </Link>
+        {date.current && (
+          <button
+            className={`text-white bg-[#1e3a89] p-2 rounded-md self-center ${
+              userRequests.length < 2 && "bg-gray-300"
+            }`}
+            disabled={userRequests.length < 2}
+            onClick={() => setIsRouteOptimizedModalOpen(true)}
+          >
+            Optimize Route
+          </button>
+        )}
       </div>
       {isSearching && (
         <button
@@ -303,7 +276,7 @@ export default function AdminDashboard() {
         </button>
       )}
 
-      {!date && (
+      {!date.current && (
         <>
           {userRequests.length === 0 ? (
             <div className="text-center text-gray-500 py-4">
@@ -385,11 +358,11 @@ export default function AdminDashboard() {
         </>
       )}
 
-      {date && (
+      {date.current && (
         <div className="space-y-8">
           {userRequests.length === 0 ? (
             <div className="text-center text-gray-500 py-4">
-              No requests found for {formatDate(date)}.
+              No requests found for {formatDate(date.current!)}.
             </div>
           ) : (
             <>
@@ -452,14 +425,13 @@ export default function AdminDashboard() {
                 />
               </div>
               {/*  I see no data assigned to this part  */}
-
-              {/* <div>
+              <div>
                 <strong>Assigned Time Slot:</strong>{" "}
                 <FormattedTimeSlotDateTime
                   start={selectedRequest.assignedTimeSlot?.startTime}
                   end={selectedRequest.assignedTimeSlot?.endTime}
                 />
-              </div> */}
+              </div>
             </div>
             <div className="flex justify-between gap-4 mt-6">
               <ActionButtons
@@ -478,125 +450,42 @@ export default function AdminDashboard() {
           </div>
         </div>
       )}
-    </div>
-  );
-}
-
-function DaySection({
-  title,
-  requests,
-  onCardClick,
-}: {
-  title: string;
-  requests: AllUserRequestsAdminGetResponse[];
-  onCardClick: (req: AllUserRequestsAdminGetResponse) => void;
-}) {
-  return (
-    <div>
-      <div className="flex items-center justify-between mb-4">
-        <h3 className="text-2xl font-bold">{title}</h3>
-        {requests.length > 1 && (
-          // only display "Optimize route" button if we have min 2 requests in that time frame
-          <button
-            className="text-white bg-[#1e3a89] p-2 rounded-md self-center"
-            onClick={() => {
-              //!!! TODO: Implement route optimization here !!!
-              console.log(`Optimize route for ${title}`);
-            }}
-          >
-            Optimize Route
-          </button>
-        )}
-      </div>
-      {requests.length === 0 ? (
-        <div className="text-gray-500">
-          No requests found in this time period.
-        </div>
-      ) : (
-        <div className="grid grid-cols-1 sm:grid-cols-3 gap-6">
-          {requests.map((request) => (
-            <div
-              key={request.id}
-              className={`rounded-lg shadow-lg p-6 hover:shadow-xl transition border-4 cursor-pointer flex flex-col ${
-                statusToColorMap[request.status].border
-              } ${statusToColorMap[request.status].bg}`}
-              onClick={() => onCardClick(request)}
-            >
-              <div className="flex justify-between w-full mb-4">
-                <div className="text-gray-600 font-medium">
-                  {request.user.name}
-                </div>
-                <div className="text-gray-600 font-medium">
-                  <FormattedTimeSlotDateTime
-                    start={request.requestedTimeSlot.startTime}
-                    end={request.requestedTimeSlot.endTime}
-                  />
-                </div>
-              </div>
-              <div className="text-gray-700 mb-2">
-                {request.address?.city || "Unknown City"},{" "}
-                {request.address?.streetName || "Unknown Street"}
-              </div>
-              <h2
-                className={`text-lg font-bold uppercase ${
-                  statusToColorMap[request.status].text
-                }`}
-              >
-                {request.status}
-              </h2>
-            </div>
-          ))}
-        </div>
+      {isRouteOptimizedModalOpen && (
+        <RouteOptimizerModal
+          onClose={() => setIsRouteOptimizedModalOpen(false)}
+          dateValue={date.current!}
+        />
       )}
     </div>
   );
 }
 
-const ActionButtons = ({
-  request,
-  onActionButtonClicked,
+const formatDate = (dateString: string | Date) => {
+  if (!dateString) return "";
+  const date =
+    typeof dateString === "string" ? new Date(dateString) : dateString;
+  return date.toLocaleDateString("en-US", {
+    month: "short",
+    day: "numeric",
+    year: "numeric",
+  });
+};
+
+const RouteOptimizerModal = ({
+  onClose,
+  dateValue,
 }: {
-  request: AllUserRequestsAdminGetResponse;
-  onActionButtonClicked: (requestStatus: RequestStatus) => void;
+  onClose: () => void;
+  dateValue: string;
 }) => {
-  const statuses: RequestStatus[] = [
-    RequestStatus.PENDING,
-    RequestStatus.COMPLETED,
-    RequestStatus.CANCELED,
-  ];
-
   return (
-    <div className="flex gap-2 flex-wrap">
-      {statuses.map((status) => {
-        let bgColor = "";
-        let hoverColor = "";
-
-        if (status === RequestStatus.COMPLETED) {
-          bgColor = "bg-green-500";
-          hoverColor = "hover:bg-green-600";
-        } else if (status === RequestStatus.PENDING) {
-          bgColor = "bg-yellow-500";
-          hoverColor = "hover:bg-yellow-600";
-        } else if (status === RequestStatus.CANCELED) {
-          bgColor = "bg-red-500";
-          hoverColor = "hover:bg-red-600";
-        }
-
-        const isCurrent = request.status === status;
-
-        return (
-          <button
-            key={status}
-            className={`${bgColor} ${hoverColor} text-white px-4 py-2 rounded transition ${
-              isCurrent ? "opacity-50 cursor-not-allowed" : ""
-            }`}
-            onClick={() => !isCurrent && onActionButtonClicked(status)}
-            disabled={isCurrent}
-          >
-            {status.charAt(0) + status.slice(1).toLowerCase()}
-          </button>
-        );
-      })}
+    <div className="absolute w-full top-[0] h-full bg-simmpy-gray-900/60">
+      <div className="modal-box w-full max-w-[800px] lg:ml-[30%] p-4 rounded-md mt-[20vh] bg-simmpy-gray-100">
+        <div className="flex justify-end">
+          <button onClick={onClose}>X</button>
+        </div>
+        <RouteManager dateValue={dateValue} />
+      </div>
     </div>
   );
 };
